requirements.txt
langchain-openai
python-dotenv
pandas
PyPDF2
numpy
scikit-learn
langchain-community